requirements.txt
fastapi
uvicorn
pymongo
requests
kafka-python
